Moved filter to utils.py and added comments
