+++ hackathon_site/hackathon_site/utils.py
@@ -1,4 +1,7 @@
 from datetime import datetime
+from dateutil import tz
+from sys import platform
+
 
 from django.conf import settings
 
@@ -15,3 +18,18 @@ def is_registration_open():
     # is configured to match TIME_ZONE. We then make the datetime object timezone-aware.
     now = datetime.now().replace(tzinfo=settings.TZ_INFO)
     return settings.REGISTRATION_OPEN_DATE <= now < settings.REGISTRATION_CLOSE_DATE
+
+
+def formatDateTime(datetime, format):
+    """
+    Format a datetime using strftime. 
+    """
+    local_zone = tz.tzlocal()
+    # Replaces incompatible identifiers.
+    # Add to this function if there are any datetime format issues
+    if platform == "win32":
+        finalFormat = format.replace("%-d", "%#d")
+    else:
+        finalFormat = format.replace("%#d", "%-d")
+    retVal = datetime.astimezone(local_zone).strftime(finalFormat)
+    return retVal
